Restored Support for configurable pci bus slots for guest vnics
This commit was inadvertently omitted when setting up stx-nova.

This feature introduces the capability to specify the guest PCI
addresses for network interfaces.

The REST API has been extended to support the new parameter
wrs-if:vif_pci_address when requesting a nic for a particular network.
The format of the PCI address is: <domain>:<bus>:<slot>.<function>.
However in this release it's not possible to specify a domain value
other than 0 (because of limitation in kvm/qemu) and a function value
other than 0 (this add complexity in the management of PCI devices for
the libvirt XML specification).

The possibility to specify the bus and the slot allows the user to
address a wide range of devices within it's guest (bus is a 16 bits
value and slot is a 5 bits value). If a bus number greater than 0 is
specified, the corresponding pci-bridges will be configured in the
libvirt XML.

The virtual PCI address is persisted in the associated Neutron port in
the profile field.

Additionaly, set pci controller pcihole64 attribute to 64 GiB: This
looks at the os_type of an instance, and if it is windows, then the
pcihole64 attribute is set to 64 GiB.

In the case of windows guests however, if we allocate this space, then
the guest will fail to boot.  As such, we discriminate based on os_type,
and do not allocate this in the case of windows guests.

Limitations:
- This feature only support PCI devices of type NIC. Other PCI devices
  that can be specified with a flavor is not supported.

Ported from internal commit bf5593bb752f59edcec7ea9c0569e1952eb6155d

StoryBoard:2002877 Task:22839
https://storyboard.openstack.org/#!/story/2002877