Fix configurable pci bus slots for PCI passthrough vnics
Commit 7020c70 ported an internal commit for allowing configurable
PCI bus slots for vnics, but did not include a subsequent internal
commit that removed the modifications to _guest_add_pci_devices.

As such, the commit relies on an obsolete PciDevice object, which
causes an exception to occur in stx-nova for any instance trying
to boot with PCI passthrough vnics.

In fact, we can return _guest_add_pci_devices to it's upstream
signature and behaviour, as the configurable address is set in
the general _get_guest_config function.

Testing:

- Launched instances with PCI passthrough vnics. Ensured
vif PCI address could be set via the boot command.
- Tox

Story: 2002877
Task: 22839

Closes-Bug: #1796952
Closes-Bug: #1796198