Conditionally map compute hosts when created
Map compute hosts only if map_new_hosts is set to True.
This is needed to remove the current StarlingX nova extension that
allows nova services to be created before the service is actually
running.
Cosmetic change to libvirt/config.py to allow pep8 to succeed
This is a temporary change that will no longer be needed when stx
moves to Stein.

Story: 2004583
Task: 28387
Signed-off-by: Daniel Chavolla <daniel.chavolla@windriver.com>