[BP] LM: add support for updating NUMA-related XML on the source
With this change, the source checks for the presence of
dst_numa_config field in the migrate_data that it receives. If
present, the source will update the XML it sends to the dest with new
CPU pinning, cell pinning, and/or emulator thread pinning information.

Backported from https://review.opendev.org/#/c/635229/30
* no changes required
* a unit test problem is present in 635229 that is addressed
  in the following commit, kept separate to simplify sending back upstream

Implements blueprint bp/numa-aware-live-migration
Change-Id: I0f40f0461c34c49c6e26dcae5e9bd78ee7f1a720